Basic internal implementation of BamFile & BamPipe  * BgzfStream now working on IBamIODevice instead of FILE*  * BamReaderPrivate now queries stream's IsOpen() method instead of touching member variable directly  * Empty implementations of BamHttp & BamFtp  * Added global BT_ASSERT_X macro for convenience

+++ src/api/internal/BamHeader_p.cpp
@@ -32,7 +32,7 @@ bool BamHeader::CheckMagicNumber(BgzfStream* stream) {
 
     // try to read magic number
     char buffer[Constants::BAM_HEADER_MAGIC_LENGTH];
-    if ( stream->Read(buffer, Constants::BAM_HEADER_MAGIC_LENGTH) != (int)Constants::BAM_HEADER_MAGIC_LENGTH ) {
+    if ( stream->Read(buffer, Constants::BAM_HEADER_MAGIC_LENGTH) != Constants::BAM_HEADER_MAGIC_LENGTH ) {
         fprintf(stderr, "BamHeader ERROR: could not read magic number\n");
         return false;
     }
